Improve example in README.md
Use an absolute rather than a relative import, as this is required when users install
from npm.

We also add a full example, because the API is slightly different from tape. Tape
requires users to load the test function, or instantiate an instance, whereas
timed-tape does the instantiation for the user.